bug fix for gcc3.4 tazz
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@2448 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Framework/FWAppGLUT.cpp
@@ -32,7 +32,7 @@ void FWAppGLUT::GlutTimerFunc(int id){
 	PHSceneIf* ps = fs->GetPHScene();
 	if(!ps)return;
 
-	int timeStep = ps->GetTimeStep() * 1000.0;
+	int timeStep = (int)(ps->GetTimeStep() * 1000.0);
 	if (timeStep<1) timeStep = 1;
 	glutTimerFunc(timeStep, GlutTimerFunc, 0);
 }
